backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
anthropic==0.18.0
openai==1.10.0
pandas==2.1.4
python-dotenv==1.0.0
pydantic==2.5.3
